+++ core/services/fluxmonitorv2/flux_monitor_test.go
@@ -90,10 +90,11 @@ type testMocks struct {
 	pipelineRunner    *pipelinemocks.Runner
 	keyStore          *fmmocks.KeyStoreInterface
 	contractSubmitter *fmmocks.ContractSubmitter
+	flags             *fmmocks.Flags
 }
 
-func newTestMocks() *testMocks {
-	return &testMocks{
+func newTestMocks(t *testing.T) *testMocks {
+	tm := &testMocks{
 		fluxAggregator:    new(mocks.FluxAggregator),
 		logBroadcast:      new(logmocks.Broadcast),
 		logBroadcaster:    new(logmocks.Broadcaster),
@@ -103,7 +104,24 @@ func newTestMocks() *testMocks {
 		pipelineRunner:    new(pipelinemocks.Runner),
 		keyStore:          new(fmmocks.KeyStoreInterface),
 		contractSubmitter: new(fmmocks.ContractSubmitter),
+		flags:             new(fmmocks.Flags),
 	}
+
+	tm.flags.On("ContractExists").Maybe().Return(false)
+	tm.logBroadcast.On("String").Maybe().Return("")
+
+	tm.fluxAggregator.Test(t)
+	tm.logBroadcast.Test(t)
+	tm.logBroadcaster.Test(t)
+	tm.orm.Test(t)
+	tm.jobORM.Test(t)
+	tm.pipelineORM.Test(t)
+	tm.pipelineRunner.Test(t)
+	tm.keyStore.Test(t)
+	tm.contractSubmitter.Test(t)
+	tm.flags.Test(t)
+
+	return tm
 }
 
 // AssertExpectations asserts expectations of all the mocks
@@ -117,28 +135,27 @@ func (tm *testMocks) AssertExpectations(t *testing.T) {
 	tm.pipelineRunner.AssertExpectations(t)
 	tm.keyStore.AssertExpectations(t)
 	tm.contractSubmitter.AssertExpectations(t)
+	tm.flags.AssertExpectations(t)
 }
 
 func setupMocks(t *testing.T) *testMocks {
 	t.Helper()
 
-	tm := newTestMocks()
-
-	t.Cleanup(func() {
-		tm.AssertExpectations(t)
-	})
+	tm := newTestMocks(t)
 
 	return tm
 }
 
 type setupOptions struct {
-	pollTickerDisabled  bool
-	idleTimerDisabled   bool
-	idleTimerPeriod     time.Duration
-	drumbeatEnabled     bool
-	drumbeatSchedule    string
-	drumbeatRandomDelay time.Duration
-	orm                 fluxmonitorv2.ORM
+	pollTickerDisabled    bool
+	idleTimerDisabled     bool
+	idleTimerPeriod       time.Duration
+	drumbeatEnabled       bool
+	drumbeatSchedule      string
+	drumbeatRandomDelay   time.Duration
+	hibernationPollPeriod time.Duration
+	flags                 *fmmocks.Flags
+	orm                   fluxmonitorv2.ORM
 }
 
 // setup sets up a Flux Monitor for testing, allowing the test to provide
@@ -148,13 +165,17 @@ func setup(t *testing.T, db *gorm.DB, optionFns ...func(*setupOptions)) (*fluxmo
 
 	tm := setupMocks(t)
 	options := setupOptions{
-		idleTimerPeriod: time.Minute,
-		orm:             tm.orm,
+		idleTimerPeriod:       time.Minute,
+		hibernationPollPeriod: fluxmonitorv2.DefaultHibernationPollPeriod,
+		flags:                 tm.flags,
+		orm:                   tm.orm,
 	}
 
 	for _, optionFn := range optionFns {
 		optionFn(&options)
 	}
+
+	tm.flags = options.flags
 
 	pollManager, err := fluxmonitorv2.NewPollManager(
 		fluxmonitorv2.PollManagerConfig{
@@ -165,7 +186,7 @@ func setup(t *testing.T, db *gorm.DB, optionFns ...func(*setupOptions)) (*fluxmo
 			DrumbeatEnabled:         options.drumbeatEnabled,
 			DrumbeatSchedule:        options.drumbeatSchedule,
 			DrumbeatRandomDelay:     options.drumbeatRandomDelay,
-			HibernationPollPeriod:   24 * time.Hour,
+			HibernationPollPeriod:   options.hibernationPollPeriod,
 			MinRetryBackoffDuration: 1 * time.Minute,
 			MaxRetryBackoffDuration: 1 * time.Hour,
 		},
@@ -188,7 +209,7 @@ func setup(t *testing.T, db *gorm.DB, optionFns ...func(*setupOptions)) (*fluxmo
 		tm.contractSubmitter,
 		fluxmonitorv2.NewDeviationChecker(threshold, absoluteThreshold),
 		fluxmonitorv2.NewSubmissionChecker(big.NewInt(0), big.NewInt(100000000000)),
-		fluxmonitorv2.Flags{},
+		options.flags,
 		tm.fluxAggregator,
 		tm.logBroadcaster,
 		logger.Default,
@@ -225,6 +246,23 @@ func enableDrumbeatTicker(schedule string, randomDelay time.Duration) func(*setu
 func setIdleTimerPeriod(period time.Duration) func(*setupOptions) {
 	return func(opts *setupOptions) {
 		opts.idleTimerPeriod = period
+	}
+}
+
+// setHibernationTickerPeriod is an option to set the hibernation ticker period during setup
+func setHibernationTickerPeriod(period time.Duration) func(*setupOptions) {
+	return func(opts *setupOptions) {
+		opts.hibernationPollPeriod = period
+	}
+}
+
+// setHibernationTickerPeriod is an option to set the hibernation ticker period during setup
+func setHibernationState(hibernating bool) func(*setupOptions) {
+	return func(opts *setupOptions) {
+		opts.flags = new(fmmocks.Flags)
+		opts.flags.On("ContractExists").Return(true)
+		opts.flags.On("Address").Return(common.Address{})
+		opts.flags.On("IsLowered", mock.Anything).Return(!hibernating, nil)
 	}
 }
 
@@ -443,7 +481,7 @@ func TestFluxMonitor_PollIfEligible(t *testing.T) {
 			tm.fluxAggregator.On("GetOracles", nilOpts).Return(oracles, nil)
 			fm.SetOracleAddress()
 			fm.ExportedPollIfEligible(thresholds.rel, thresholds.abs)
-			tm.logBroadcaster.AssertExpectations(t)
+			tm.AssertExpectations(t)
 		})
 	}
 }
@@ -479,7 +517,7 @@ func TestFluxMonitor_PollIfEligible_Creates_JobErr(t *testing.T) {
 	require.NoError(t, fm.SetOracleAddress())
 
 	fm.ExportedPollIfEligible(1, 1)
-	tm.logBroadcaster.AssertExpectations(t)
+	tm.AssertExpectations(t)
 }
 
 func TestPollingDeviationChecker_BuffersLogs(t *testing.T) {
@@ -649,6 +687,7 @@ func TestPollingDeviationChecker_BuffersLogs(t *testing.T) {
 	for i := 1; i <= 4; i++ {
 		logBroadcast := new(logmocks.Broadcast)
 		logBroadcast.On("DecodedLog").Return(&flux_aggregator_wrapper.FluxAggregatorNewRound{RoundId: big.NewInt(int64(i)), StartedAt: big.NewInt(0)})
+		logBroadcast.On("String").Maybe().Return("")
 		tm.logBroadcaster.On("WasAlreadyConsumed", mock.Anything, mock.Anything).Return(false, nil)
 		tm.logBroadcaster.On("MarkConsumed", mock.Anything, mock.Anything).Return(nil)
 		logBroadcasts = append(logBroadcasts, logBroadcast)
@@ -663,6 +702,7 @@ func TestPollingDeviationChecker_BuffersLogs(t *testing.T) {
 	close(chBlock)
 	<-chSafeToAssert
 	fm.Close()
+	tm.AssertExpectations(t)
 }
 
 func TestFluxMonitor_TriggerIdleTimeThreshold(t *testing.T) {
@@ -749,8 +789,94 @@ func TestFluxMonitor_TriggerIdleTimeThreshold(t *testing.T) {
 			if !tc.expectedToSubmit {
 				require.Len(t, idleDurationOccured, 0)
 			}
+			tm.AssertExpectations(t)
 		})
 	}
+}
+
+func TestFluxMonitor_HibernationTickerFiresMultipleTimes(t *testing.T) {
+	t.Parallel()
+
+	store, nodeAddr := setupStoreWithKey(t)
+	oracles := []common.Address{nodeAddr, cltest.NewAddress()}
+
+	fm, tm := setup(t,
+		store.DB,
+		disablePollTicker(true),
+		disableIdleTimer(true),
+		setHibernationTickerPeriod(time.Second),
+		setHibernationState(true),
+	)
+
+	tm.keyStore.On("SendingKeys").Return([]ethkey.Key{{Address: ethkey.EIP55AddressFromAddress(nodeAddr)}}, nil).Once()
+
+	const fetchedAnswer = 100
+	answerBigInt := big.NewInt(fetchedAnswer)
+
+	tm.fluxAggregator.On("Address").Return(contractAddress)
+	tm.fluxAggregator.On("GetOracles", nilOpts).Return(oracles, nil)
+	tm.logBroadcaster.On("Register", mock.Anything, mock.Anything).Return(func() {})
+	tm.logBroadcaster.On("IsConnected").Return(true)
+	tm.fluxAggregator.On("LatestRoundData", nilOpts).Return(freshContractRoundDataResponse()).Once()
+
+	pollOccured := make(chan struct{}, 4)
+
+	err := fm.Start()
+	require.NoError(t, err)
+
+	t.Cleanup(func() { fm.Close() })
+
+	roundState1 := flux_aggregator_wrapper.OracleRoundState{RoundId: 1, EligibleToSubmit: false, LatestSubmission: answerBigInt, StartedAt: now()}
+	tm.fluxAggregator.On("OracleRoundState", nilOpts, nodeAddr, uint32(0)).Return(roundState1, nil).Once().Run(func(args mock.Arguments) {
+		pollOccured <- struct{}{}
+	})
+	tm.orm.
+		On("FindOrCreateFluxMonitorRoundStats", contractAddress, uint32(1)).
+		Return(fluxmonitorv2.FluxMonitorRoundStatsV2{
+			Aggregator:     contractAddress,
+			RoundID:        1,
+			NumSubmissions: 0,
+		}, nil).Once()
+
+	require.Eventually(t, func() bool { return len(pollOccured) == 1 }, 3*time.Second, 10*time.Millisecond)
+
+	// hiberation tick 1 triggers using the same round id as the initial poll. This resets the idle timer
+	roundState1Responded := flux_aggregator_wrapper.OracleRoundState{RoundId: 1, EligibleToSubmit: false, LatestSubmission: answerBigInt, StartedAt: now() + 1}
+	tm.fluxAggregator.On("OracleRoundState", nilOpts, nodeAddr, uint32(0)).Return(roundState1Responded, nil).Once().Run(func(args mock.Arguments) {
+		pollOccured <- struct{}{}
+	})
+
+	// Finds an existing run created by the initial poll
+	tm.orm.
+		On("FindOrCreateFluxMonitorRoundStats", contractAddress, uint32(1)).
+		Return(fluxmonitorv2.FluxMonitorRoundStatsV2{
+			PipelineRunID:  corenull.NewInt64(int64(1), true),
+			Aggregator:     contractAddress,
+			RoundID:        1,
+			NumSubmissions: 1,
+		}, nil).Once()
+	finishedAt := time.Now()
+	tm.pipelineORM.On("FindRun", int64(1)).Return(pipeline.Run{
+		FinishedAt: null.TimeFrom(finishedAt),
+	}, nil)
+
+	require.Eventually(t, func() bool { return len(pollOccured) == 2 }, 6*time.Second, 10*time.Millisecond)
+
+	// hiberation tick 2 triggers a new round. Started at is 0
+	roundState2 := flux_aggregator_wrapper.OracleRoundState{RoundId: 2, EligibleToSubmit: false, LatestSubmission: answerBigInt, StartedAt: 0}
+	tm.fluxAggregator.On("OracleRoundState", nilOpts, nodeAddr, uint32(0)).Return(roundState2, nil).Once().Run(func(args mock.Arguments) {
+		pollOccured <- struct{}{}
+	})
+	tm.orm.
+		On("FindOrCreateFluxMonitorRoundStats", contractAddress, uint32(2)).
+		Return(fluxmonitorv2.FluxMonitorRoundStatsV2{
+			Aggregator:     contractAddress,
+			RoundID:        2,
+			NumSubmissions: 0,
+		}, nil).Once()
+
+	require.Eventually(t, func() bool { return len(pollOccured) == 3 }, 8*time.Second, 10*time.Millisecond)
+	tm.AssertExpectations(t)
 }
 
 func TestFluxMonitor_IdleTimerResetsOnNewRound(t *testing.T) {
@@ -855,6 +981,7 @@ func TestFluxMonitor_IdleTimerResetsOnNewRound(t *testing.T) {
 	fm.ExportedProcessLogs()
 
 	require.Eventually(t, func() bool { return len(idleDurationOccured) == 4 }, 3*time.Second, 10*time.Millisecond)
+	tm.AssertExpectations(t)
 }
 
 func TestFluxMonitor_RoundTimeoutCausesPoll_timesOutAtZero(t *testing.T) {
@@ -903,7 +1030,7 @@ func TestFluxMonitor_RoundTimeoutCausesPoll_timesOutAtZero(t *testing.T) {
 	gomega.NewGomegaWithT(t).Eventually(ch).Should(gomega.BeClosed())
 
 	fm.Close()
-	tm.logBroadcaster.AssertExpectations(t)
+	tm.AssertExpectations(t)
 }
 
 func TestFluxMonitor_UsesPreviousRoundStateOnStartup_RoundTimeout(t *testing.T) {
@@ -965,6 +1092,7 @@ func TestFluxMonitor_UsesPreviousRoundStateOnStartup_RoundTimeout(t *testing.T) 
 			}
 
 			fm.Close()
+			tm.AssertExpectations(t)
 		})
 	}
 }
@@ -1045,6 +1173,7 @@ func TestFluxMonitor_UsesPreviousRoundStateOnStartup_IdleTimer(t *testing.T) {
 			} else {
 				gomega.NewGomegaWithT(t).Consistently(chRoundState).ShouldNot(gomega.BeClosed())
 			}
+			tm.AssertExpectations(t)
 		})
 	}
 }
@@ -1125,6 +1254,7 @@ func TestFluxMonitor_RoundTimeoutCausesPoll_timesOutNotZero(t *testing.T) {
 
 	time.Sleep(time.Duration(2*timeout) * time.Second)
 	fm.Close()
+	tm.AssertExpectations(t)
 }
 
 func TestFluxMonitor_HandlesNilLogs(t *testing.T) {
@@ -1132,13 +1262,14 @@ func TestFluxMonitor_HandlesNilLogs(t *testing.T) {
 
 	store, cleanup := cltest.NewStore(t)
 	t.Cleanup(cleanup)
-	fm, _ := setup(t, store.DB)
+	fm, tm := setup(t, store.DB)
 
 	logBroadcast := new(logmocks.Broadcast)
 	var logNewRound *flux_aggregator_wrapper.FluxAggregatorNewRound
 	var logAnswerUpdated *flux_aggregator_wrapper.FluxAggregatorAnswerUpdated
 	var randomType interface{}
 
+	logBroadcast.On("String").Maybe().Return("")
 	logBroadcast.On("DecodedLog").Return(logNewRound).Once()
 	assert.NotPanics(t, func() {
 		fm.HandleLog(logBroadcast)
@@ -1154,6 +1285,7 @@ func TestFluxMonitor_HandlesNilLogs(t *testing.T) {
 		fm.HandleLog(logBroadcast)
 	})
 
+	tm.AssertExpectations(t)
 	logBroadcast.AssertExpectations(t)
 }
 
@@ -1175,6 +1307,7 @@ func TestFluxMonitor_ConsumeLogBroadcast(t *testing.T) {
 
 	fm.ExportedBacklog().Add(fluxmonitorv2.PriorityNewRoundLog, tm.logBroadcast)
 	fm.ExportedProcessLogs()
+	tm.AssertExpectations(t)
 }
 
 func TestFluxMonitor_ConsumeLogBroadcast_Error(t *testing.T) {
@@ -1202,6 +1335,7 @@ func TestFluxMonitor_ConsumeLogBroadcast_Error(t *testing.T) {
 
 			fm.ExportedBacklog().Add(fluxmonitorv2.PriorityNewRoundLog, tm.logBroadcast)
 			fm.ExportedProcessLogs()
+			tm.AssertExpectations(t)
 		})
 	}
 }
@@ -1313,6 +1447,7 @@ func TestFluxMonitor_DoesNotDoubleSubmit(t *testing.T) {
 		}, nil)
 
 		fm.ExportedPollIfEligible(0, 0)
+		tm.AssertExpectations(t)
 	})
 
 	t.Run("when poll ticker fires, then NewRound log arrives", func(t *testing.T) {
@@ -1403,6 +1538,7 @@ func TestFluxMonitor_DoesNotDoubleSubmit(t *testing.T) {
 			RoundId:   big.NewInt(roundID),
 			StartedAt: big.NewInt(0),
 		}, log.NewLogBroadcast(types.Log{}, nil))
+		tm.AssertExpectations(t)
 	})
 }
 
@@ -1512,4 +1648,5 @@ func TestFluxMonitor_DrumbeatTicker(t *testing.T) {
 	cltest.EventuallyExpectationsMet(t, tm.orm, waitTime, interval)
 	cltest.EventuallyExpectationsMet(t, tm.pipelineORM, waitTime, interval)
 	cltest.EventuallyExpectationsMet(t, tm.contractSubmitter, waitTime, interval)
+	tm.AssertExpectations(t)
 }