+++ docs/CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.10.12] - 2021-08-16
+
+### Fixed
+
+- Resolved FMv2 stalling in Hibernation mode
+- Resolved rare issue when the Gas Estimator fails on start
+- Resolved the handling of nil values for gas price
+
 ## [0.10.11] - 2021-08-09
 
 A new configuration variable, `BLOCK_BACKFILL_SKIP`, can be optionally set to "true" in order to strongly limit the depth of the log backfill.